Implemented a simple dynamic string for a return date
My motive in this was to allow for the initial requirement to set a "We will return by..." type of string
in the offline text.

The implementation is simple, it uses the RETURN_STRING env var and passes the index.html file through envsubst. This allows
for us to, for example, use -e RETURN_STRING="We will return on Wednesday (June 12, 2019)." when running this image in Docker. If
the ENV Var is unset, it will substitute the variable with nothing, leaving a generic, dateless message.

I added an entrypoint.sh that runs the original text through envsubst and redirects it to a temp file, then copy the date file over to index.html.
Then we run nginx -g "daemon off;"